refactor(plugin-cli): rename neko-plugin-cli → neko_plugin_cli
Drop the hyphenated directory and the matching `sys.path.insert` +
`from public import ...` runtime hack. The hyphen made the directory
unimportable as a Python package, forcing every caller to inject sys.path
and reference a generic top-level name `public`. That pattern bit Nuitka
twice in one day:

1. Nuitka cannot follow `from public import` statically, so the public
   package was never compiled into the standalone exe.
2. The compensating `--include-data-dir=plugin/neko-plugin-cli` silently
   filtered .py files (Nuitka's documented default for data dirs), so
   even with the explicit include, only docs/*.md shipped — server-side
   plugin_cli/service.py raised ModuleNotFoundError: No module named
   'public' and the embedded user plugin server failed to start, which
   in turn made the plugin management page unreachable.

Renaming to a valid Python package eliminates both layers: the package
is reachable as `plugin.neko_plugin_cli.public` and is auto-included via
`--include-package=plugin` (already in build_nuitka.bat and .github/
workflows/build-desktop.yml). The sys.path injections in cli.py,
service.py, and 5 test files are deleted; absolute/relative package
imports replace `from public import ...`.

Direct script invocation (`python plugin/neko_plugin_cli/cli.py ...`)
no longer works; use the standard module form
`python -m plugin.neko_plugin_cli.cli ...` documented in
docs/package-format.md.

The previously-added `--include-data-dir=plugin/neko-plugin-cli` line in
build_nuitka.bat is removed (no longer needed; --include-package=plugin
covers it). build-desktop.yml needs no change — it never had a
neko-plugin-cli entry and the new package name is auto-discovered.

A separate concurrent task already moved `_TARGET_ROOT` from
`<bundle>/plugin/neko-plugin-cli/target` to the user-writable
`USER_PLUGIN_PACKAGES_ROOT`; the now-misleading UI placeholder
"默认使用 neko-plugin-cli/target" is updated to a generic
"默认使用应用插件包目录".

Tests: 32 passed (plugin/tests/unit/test_neko_plugin_cli_*.py +
plugin/tests/integration/test_neko_plugin_cli_*.py +
plugin/tests/unit/server/test_plugin_cli_route.py).
Smoke: `python -m plugin.neko_plugin_cli.cli --help` works;
`from plugin.server.http_app import build_plugin_server_app` (the
exact import chain that failed at runtime in the bundled exe) now
succeeds without sys.path hacks.
