+++ plugin/neko_plugin_cli/public/bundle_analysis.py
