+++ plugin/neko_plugin_cli/public/inspect.py
