+++ plugin/neko_plugin_cli/public/pack_rules.py
