+++ plugin/neko_plugin_cli/public/plugin_source.py
